seller signup issue fixed
